pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.4.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.45.1",
    "streamlit-option-menu>=0.4.0",
    "uvicorn>=0.34.2",
]
